+++ supabase/migrations/20260201222844_enable_realtime.sql
@@ -1,0 +1,35 @@
+alter publication supabase_realtime
+add table assignments;
+
+alter publication supabase_realtime
+add table conflicts;
+
+alter publication supabase_realtime
+add table currencies;
+
+alter publication supabase_realtime
+add table proposals;
+
+alter publication supabase_realtime
+add table roles;
+
+alter publication supabase_realtime
+add table scholars;
+
+alter publication supabase_realtime
+add table submissions;
+
+alter publication supabase_realtime
+add table supporters;
+
+alter publication supabase_realtime
+add table tokens;
+
+alter publication supabase_realtime
+add table transactions;
+
+alter publication supabase_realtime
+add table venues;
+
+alter publication supabase_realtime
+add table volunteers;